WP-18 добавлена основная логика игры
